haptic feedback fix
