+++ src/senseglove/senseglove_haptics/config/HapticLevels.cfg
@@ -17,6 +17,12 @@ gen.add("middle_buzz", double_t, 0, "Vibration Feedback for the middle finger", 
 gen.add("ring_buzz",   double_t, 0, "Vibration Feedback for the ring finger",    0.0, 0.0, 100.0)
 gen.add("pinky_buzz",  double_t, 0, "Vibration Feedback for the pinky finger",   0.0, 0.0, 100.0)
 
-gen.add("reset", bool_t, 0, "A Boolean parameter to reset",  False)
+gen.add("thumper_buzz",  double_t, 0, "Vibration Feedback for the Thumper",   0.0, 0.0, 100.0)
 
-exit(gen.generate(PACKAGE, "senseglove_haptics_node", "HapticSlider"))
+gen.add("palm_index_buzz", double_t, 0, "Vibration Feedback for the palm_index side",  0.0, 0.0, 100.0)
+gen.add("palm_pinky_buzz", double_t, 0, "Vibration Feedback for the palm_pinky side",    0.0, 0.0, 100.0)
+gen.add("palm_strap",      double_t, 0, "Active Strap Squeeze",   0.0, 0.0, 100.0)
+
+gen.add("reset", bool_t, 0, "A Boolean parameter to reset", False)
+
+exit(gen.generate(PACKAGE, "senseglove_haptics_node_dynamic", "HapticSlider"))
