+++ src/senseglove/senseglove_haptics/scripts/senseglove_haptics_node
@@ -1,5 +1,7 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
+
 import senseglove_haptics.haptics_node_action_interface
+import senseglove_haptics.haptics_node_simple
 
 if __name__ == '__main__':
-    senseglove_haptics.haptics_node_action_interface.main()
+    senseglove_haptics.haptics_node_simple.main()
